refactored Events, added Events section to OS page, changed column width on events
